Update the update script.
* git/update: adl helps me in this script: The function
`file_list_from_message' is written by him, as the deletion of the
init function (replaced by map), as the way to understand
if a "file is modified by commit and not mentionned in log", and this
opposite.
I modified this script to be able to use this script as a "commit-msg"
hook. I add other lexical checks.